redo BOW_embedding
Last time I was embedding the characters, not the words; that led to the conclusion of 78%, which was lower than 86% from no embedding. That led me to conclude embeding always loses information.

Results from this run show that an embed_size of 20 gives ~85%, statistically indistinguishable from 86% from no embedding. Thus, for an embed_size of 20, all information is retained.